171. Quick bug fix- win condition
Added some additional modifiers to difficulty, including changing enemy speed and defender costs.

+++ Assets/Scripts/Defender.cs
@@ -12,7 +12,7 @@ public class Defender : MonoBehaviour
         FindObjectOfType<StarsDisplay>().AddStars(amount);
     }
 
-    public int GetStarCost()
+    public float GetStarCost()
     {
         return cost;
     }
